New logic

+++ src/lib/helper/fetcher.ts
@@ -1,5 +1,5 @@
+import { goto } from "$app/navigation"
 import type { addColumnInterface } from "$lib/helper/helper"
-
 export interface createDbInterface{
     user_table:string
     table_description:string
@@ -24,8 +24,11 @@ export async function auth(login:string,password:string) {
             })
         })
         const response = await request.json()
+        console.log(response)
+        
         if(!request.ok) throw new Error(response.error)
         localStorage.setItem("user",response.login)
+        localStorage.setItem("currentDB",response.db_list.split(",")[0].split(".")[0])
         location.reload()
     }
     catch(err){
@@ -45,12 +48,15 @@ export async function createUser(newUser: newUserInterface){
                 "login":newUser.login,
                 "name":newUser.name,
                 "password":newUser.pass,
-                role:""
+                role:"user"
             })
         })
         const response = await request.json()
         if(!request.ok) throw new Error(response.error)
+        console.log(response)
         localStorage.setItem("user",response.login)
+        localStorage.setItem("currentDB",'')
+
         location.reload()
     }
     catch(err){
@@ -60,18 +66,20 @@ export async function createUser(newUser: newUserInterface){
 
 export async function createDB(db:createDbInterface) {
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return
+        if(!currentDB) goto("/")
+
         const request  = await fetch("http://reffattest.ru:5000/user_tables",{
             method:"POST",
             headers: {
                 "Content-Type": "application/json",
             },
             body:JSON.stringify({
-                "user": user,
-                "user_table": user + "_" + db.user_table.split(" ").join("_"),
+                "user": currentDB,
+                "table_name": db.user_table,
                 "table_description": db.table_description,
-                "table_columns": "id INTEGER PRIMARY KEY"
             })
         })
         const response = await request.json()
@@ -84,9 +92,12 @@ export async function createDB(db:createDbInterface) {
 
 export async function deleteDB(tableName:string) {
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
-        const request  = await fetch(`http://reffattest.ru:5000/user_tables/${user}/${tableName}`,{
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return
+        if(!currentDB) goto("/")
+
+        const request  = await fetch(`http://reffattest.ru:5000/tables/${currentDB}/${tableName}`,{
             method:"DELETE",
             headers: {
                 "Content-Type": "application/json",
@@ -103,13 +114,21 @@ export async function deleteDB(tableName:string) {
 
 export async function deleteTableRow(tableName:string, id:number | string){
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
-        const request  = await fetch(`http://reffattest.ru:5000/delete_data/${user}/${tableName}/${id}`,{
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return
+        if(!currentDB) goto("/")
+
+        const request  = await fetch(`http://reffattest.ru:5000/delete_data`,{
             method:"DELETE",
             headers: {
                 "Content-Type": "application/json",
             },
+            body:JSON.stringify({
+                db_name:currentDB,
+                table_name:tableName,
+                id:id
+            })
         })
         const response = await request.json()
         if(!request.ok) throw new Error(response.error)
@@ -122,14 +141,21 @@ export async function deleteTableRow(tableName:string, id:number | string){
 
 export async function addTableItem(tableName:string, data:object){
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
-        const request  = await fetch(`http://reffattest.ru:5000/add_data/${user}/${tableName}`,{
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return
+        if(!currentDB) goto("/")
+
+        const request  = await fetch(`http://reffattest.ru:5000/add_data`,{
             method:"POST",
             headers: {
                 "Content-Type": "application/json",
             },
-            body:JSON.stringify(data)
+            body:JSON.stringify({
+                table_name:tableName,
+                db_name:currentDB,
+                data:data
+            })
         })
         const response = await request.json()
         if(!request.ok) throw new Error(response.error)
@@ -143,14 +169,21 @@ export async function addTableItem(tableName:string, data:object){
 
 export async function deleteTableColumn(tableName:string, columnName: string){
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
-        const request  = await fetch(`http://reffattest.ru:5000/drop_columns/${user}/${tableName}`,{
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return
+        if(!currentDB) goto("/")
+
+        const request  = await fetch(`http://reffattest.ru:5000/drop_column/${currentDB}/${tableName}/${columnName.split("_")[0]}`,{
             method:"DELETE",
             headers: {
                 "Content-Type": "application/json",
-            },
-            body:JSON.stringify({"columns": [columnName.split("_")[0]]})
+            }
+            // body:JSON.stringify({
+            //     db_name:currentDB,
+            //     table_name:tableName,
+            //     column_name:columnName
+            // })
         })
         const response = await request.json()
         if(!request.ok) throw new Error(response.error)
@@ -163,14 +196,20 @@ export async function deleteTableColumn(tableName:string, columnName: string){
 
 export async function addTableColumn(tableName:string, columnName: addColumnInterface[]){
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
-        const request  = await fetch(`http://reffattest.ru:5000/add_columns/${user}/${tableName}`,{
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB)return
+        if(!currentDB) goto("/")
+
+        const request  = await fetch(`http://reffattest.ru:5000/add_columns`,{
             method:"POST",
             headers: {
                 "Content-Type": "application/json",
             },
-            body:JSON.stringify({"columns": columnName})
+            body:JSON.stringify({
+                db_name:currentDB,
+                table_name:tableName,
+                "columns": columnName})
         })
         const response = await request.json()
         if(!request.ok) throw new Error(response.error)
@@ -184,14 +223,23 @@ export async function editTableRow(tableName:string, rowValue:{[key:string]:stri
     const id = rowValue.id??-1
     if(rowValue.id) delete rowValue.id
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
-        const request  = await fetch(`http://reffattest.ru:5000/update_data/${user}/${tableName}/${id}`,{
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return
+        if(!currentDB) goto("/")
+        
+        const editRow={
+            db_name:currentDB,
+            table_name:tableName,
+            id:id,
+            new_data:rowValue
+        }
+        const request  = await fetch(`http://reffattest.ru:5000/update_data`,{
             method:"PUT",
             headers: {
                 "Content-Type": "application/json",
             },
-            body:JSON.stringify(rowValue)
+            body:JSON.stringify(editRow)
         })
         const response = await request.json()
         if(!request.ok) throw new Error(response.error)
@@ -204,13 +252,22 @@ export async function editTableRow(tableName:string, rowValue:{[key:string]:stri
 
 export async function searchTable(tableName:string,searchRow:{value:string,field:string}):Promise<{[key:string]:string}[]>{
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
-        const request  = await fetch(`http://reffattest.ru:5000/search-table/${tableName}?column=${searchRow.field}&value=${searchRow.value}`,{
-            method:"GET",
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return[]
+        if(!currentDB) goto("/")
+
+        const request  = await fetch(`http://reffattest.ru:5000/search-table`,{
+            method:"POST",
             headers: {
                 "Content-Type": "application/json",
             },
+            body:JSON.stringify({
+                db_name:currentDB,
+                table_name:tableName,
+                column:searchRow.field,
+                value:searchRow.value
+            })
         })
         const response = await request.json()
         if(!request.ok) throw new Error(response.error)
@@ -225,6 +282,7 @@ export async function searchTable(tableName:string,searchRow:{value:string,field
 
 export interface filterValueInterface{
     table:string,
+    db_name:string,
     filters:({
         column:string,
         from:string,
@@ -236,8 +294,12 @@ export interface filterValueInterface{
 }
 export async function searchFilter(filterValue:filterValueInterface):Promise<{[key:string]:string}[]>{
     try{
-        const user = localStorage.getItem('user')
-        if(!user) location.reload()
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return[]
+        if(!currentDB) goto("/")
+
+        if(currentDB) filterValue.db_name = currentDB
         const request  = await fetch(`http://reffattest.ru:5000/filter`,{
             method:"POST",
             headers: {
@@ -255,14 +317,44 @@ export async function searchFilter(filterValue:filterValueInterface):Promise<{[k
     return [] 
 }
 
-export async function downLoadTable(tableName:string){
+
+export async function renameColumnFetch(tableName:string,oldColumn:string,newColumn:string){
+    try{
+        const currentDB = localStorage.getItem('currentDB')
+        // if(!currentDB) location.reload()
+        // if(!currentDB) return
+        if(!currentDB) goto("/")
+
+        const request  = await fetch(`http://reffattest.ru:5000/rename_column`,{
+            method:"PUT",
+            headers: {
+                "Content-Type": "application/json",
+            },
+            body:JSON.stringify({
+                db_name:currentDB,
+                table_name:tableName,
+                old_column_name:oldColumn,
+                new_column_name:newColumn
+            })
+        })
+        const response = await request.json()
+        if(!request.ok) throw new Error(response.error)
+        return response
+    }
+    catch(err){
+        console.error(err)
+    }
+    return [] 
+}
+
+export async function downLoadTable(db_name:string, tableName:string){
     try{
         const response = await fetch('http://reffattest.ru:5000/export', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
             },
-            body: JSON.stringify({ table: tableName })
+            body: JSON.stringify({ table: tableName,db_name:db_name })
         });
 
         if (response.ok) {
@@ -282,4 +374,66 @@ export async function downLoadTable(tableName:string){
     catch(err){
         console.error(err)
     }
+}
+
+
+
+export async function createNewDB(newDb:string){
+    try{
+        const user = localStorage.getItem('user')
+        if(!user) location.reload()
+        const request  = await fetch(`http://reffattest.ru:5000/create_db`,{
+            method:"POST",
+            headers: {
+                "Content-Type": "application/json",
+            },
+            body:JSON.stringify({
+                login: user,
+                db_name:newDb
+            })
+        })
+        const response = await request.json()
+        if(!request.ok) throw new Error(response.error)
+        return response
+    }
+    catch(err){
+        console.error(err)
+    }
+    return
+}
+
+
+export async function deleteDbFetch(dBName:string){
+    try{
+        const user = localStorage.getItem('user')
+        if(!user) location.reload()
+        const request  = await fetch(`http://reffattest.ru:5000/delete_db/${user}/${dBName}`,{
+            method:"DELETE",
+            headers: {
+                "Content-Type": "application/json",
+            }
+        })
+        const response = await request.json()
+        if(!request.ok) throw new Error(response.error)
+        return response
+    }
+    catch(err){
+        console.error(err)
+    }
+    return
+}
+
+
+export async function downloadFile(db_name:string) {
+    fetch(`http://reffattest.ru:5000/download/${db_name}`)
+    .then(response => response.blob())
+    .then(blob => {
+        const url = URL.createObjectURL(blob);
+        const a = document.createElement('a');
+        a.href = url;
+        a.download = `${db_name}.db`;
+        a.click();
+        URL.revokeObjectURL(url);
+    })
+    .catch(error => console.error(error));
 }
